rename dataset search bar data hook file
