+++ static/app/views/dashboards/globalFilter/addFilter.spec.tsx
@@ -4,7 +4,7 @@ import type {TagCollection} from 'sentry/types/group';
 import {FieldKind} from 'sentry/utils/fields';
 import {type SearchBarData} from 'sentry/views/dashboards/datasetConfig/base';
 import AddFilter, {DATASET_CHOICES} from 'sentry/views/dashboards/globalFilter/addFilter';
-import {type DatasetSearchBarData} from 'sentry/views/dashboards/hooks/useSearchBarData';
+import {type DatasetSearchBarData} from 'sentry/views/dashboards/hooks/useDatasetSearchBarData';
 import {WidgetType} from 'sentry/views/dashboards/types';
 
 // Mock getDatasetConfig
